initial motor driver, compass improvements
